added codes

+++ e-commerce-web-api/src/routes/categories/categories.controller.js
@@ -4,9 +4,11 @@ const Category = require('../../models/categories.model');
 
 async function httpGetCategories(req, res) {
     try {
-        const categories = await Category.find({}, { '__v': 0, });
+        const categories = await Category.find({})
+            .select('name icon color')
+            .sort('+name')
 
-        if (categories.length === 0) return res.status(404).json({ error: 'No category found' });
+        if (categories.length === 0) return res.status(404).json({ message: 'No category found' });
 
         res.status(200).json({ categories });
     } catch (error) {
@@ -17,14 +19,16 @@ async function httpGetCategories(req, res) {
 async function httpGetCategoryByID(req, res) {
     try {
         if (!mongoose.isValidObjectId(req.params.id))
-            return res.status(400).send('Invalid Category Id');
+            return res.status(400).send({ message: 'Invalid Category Id' });
 
         const category = await Category.findById(req.params.id);
 
         if (!category) return res.status(404)
-            .json({ error: `Category not found` });
+            .json({ message: `Category not found` });
+        
+        const { name, icon, color, createdAt, updatedAt } = category;
 
-        res.status(200).json({ category });
+        res.status(200).json({ name, icon, color, createdAt, updatedAt });
     } catch (error) {
         res.status(500).json({ error: error.message });
     }
@@ -32,15 +36,15 @@ async function httpGetCategoryByID(req, res) {
 
 async function httpPostCategory(req, res) {
     try {
-        const newCreatedcategory = await Category.create(
+        const newCreatedCategory = await Category.create(
             {
                 name: req.body.name,
                 icon: req.body.icon,
                 color: req.body.color
             }
         )
-
-        res.status(201).json({ newCreatedcategory });
+        const { _id, name, icon, color } = newCreatedCategory;
+        res.status(201).json({ _id, name, icon, color });
     } catch (error) {
         res.status(500).json({ error: error.message });
     }
@@ -48,9 +52,12 @@ async function httpPostCategory(req, res) {
 
 async function httpUpdateCategory(req, res) {
     try {
+        if (!mongoose.isValidObjectId(req.params.id))
+            return res.status(400).send({ message: 'Invalid Category Id' });
+        
         const id = await Category.findById(req.params.id);
 
-        if (!id) return res.status(404).json({ error: 'Category not found' });
+        if (!id) return res.status(404).json({ message: 'Category not found' });
 
         const category = await Category.findByIdAndUpdate(
             req.params.id,
@@ -64,7 +71,8 @@ async function httpUpdateCategory(req, res) {
 
         const updatedCategory = await category.save();
 
-        res.status(200).json({ updatedCategory });
+        const { name, icon, color, updatedAt } = updatedCategory;
+        res.status(200).json({ name, icon, color, updatedAt });
     } catch (error) {
         res.status(500).json({ error: error.message });
     }
@@ -72,9 +80,12 @@ async function httpUpdateCategory(req, res) {
 
 async function httpDeleteCategory(req, res) {
     try {
+        if (!mongoose.isValidObjectId(req.params.id))
+            return res.status(400).send({ message: 'Invalid Category Id' });
+        
         const category = await Category.deleteOne({ _id: req.params.id });
 
-        if (category.deletedCount === 0) return res.status(404).json({ error: 'Category not found' });
+        if (category.deletedCount === 0) return res.status(404).json({ message: 'Category not found' });
 
         res.status(200).json({ message: 'Category has been deleted' });
     } catch (error) {
